jetson_rc_control/jetson/requirements.txt
# Python dependencies for Jetson RC Control teleoperation
# Install with: pip install -r requirements.txt

pygame>=2.0.0
Jetson.GPIO>=2.0.0
